Backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
docopt==0.6.2
google-play-scraper==1.2.7
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
packaging==25.0
pipreqs==0.4.13
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.41
sqlmodel==0.0.24
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
yarg==0.1.10
whitenoise>=6.0.0
dj-database-url>=2.0.0